server/requirements.txt
discord.py
flask
psycopg2-binary
python-dotenv
pytz
